requirements.txt
argon2_cffi>=19.1.0
backports.tempfile==1.0
cachetools==3.0.0
elasticsearch==6.4.0
enum==0.4.6
fs==2.0.16
flywheel-common[gevent]==2.0.2
ipython==5.7.0
jsonschema==2.6.0
Markdown==2.6.11
ply==3.11
prometheus_client==0.2.0
pyjwt==1.7.1
pymongo==3.7.1
python-dateutil==2.7.3
python-dxf==7.4.1
pytz==2018.4
requests[security]>=2.21.0
rfc3987==1.3.7
strict-rfc3339==0.7
unicodecsv==0.9.0
urllib3>=1.25.2
webapp2==2.5.2
WebOb==1.8.2
git+https://github.com/flywheel-io/gears.git@v0.1.8#egg=gears
